I2C driver written and humidity sensor working. Brief speed test completed, ~100 frames of ~40 bytes in ~4 seconds, i.e. ~1kB/s but with prints at 9600 b/s which slowed it down. Need to get pressure sensor readings then ready for PCB design

+++ sensor.h
@@ -34,7 +34,7 @@
 #include <xc.h> // include processor files - each processor file is guarded.  
 
 void sensor_init(void);
-uint16_t sensor_readHumidity(void);
+uint8_t sensor_readHumidity(void);
 uint16_t sensor_readTemperature(void);
 uint16_t sensor_readLastTemperature(void);
 uint16_t sensor_readPressure(void);
